# perf(helper/cookie): fast-path for name specified