--- a/src/utils/cookie.test.ts
+++ b/src/utils/cookie.test.ts
@@ -7,6 +7,7 @@ describe('Parse cookie', () => {
     const cookie: Cookie = parse(cookieString)
     expect(cookie['yummy_cookie']).toBe('choco')
     expect(cookie['tasty_cookie']).toBe('strawberry')
+    expect(cookie['no_such_cookie']).toBeUndefined()
   })
 
   it('Should parse quoted cookie values', () => {
@@ -37,6 +38,7 @@ describe('Parse cookie', () => {
     expect(cookie['tasty_cookie']).toBe('')
     expect(cookie['best_cookie']).toBe('')
     expect(cookie['last_cookie']).toBe('')
+    expect(cookie['no_such_cookie']).toBeUndefined()
   })
 
   it('Should parse cookies but not process signed cookies', () => {
@@ -48,6 +50,7 @@ describe('Parse cookie', () => {
     expect(cookie['tasty_cookie']).toBe('strawberry.I9qAeGQOvWjCEJgRPmrw90JjYpnnX2C9zoOiGSxh1Ig=')
     expect(cookie['great_cookie']).toBe('rating3.5')
     expect(cookie['best_cookie']).toBe('sugar.valueShapedLikeASignatureButIsNotASignature=')
+    expect(cookie['no_such_cookie']).toBeUndefined()
   })
 
   it('Should ignore invalid cookie names', () => {